Initial Commit
This initial version is a simple setup that implements the data I found about the format, I wanted to at least see what I could do initially.

This code is currently only set up for the macOS Catalina and later variant of the Music Library format. The earlier format that iTunes uses appears to be completely different, from what the documentation I found mentions about it.